Split FreeAgentFixture class, add InvoiceList JSON
